backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic-settings>=2.0.3
websockets>=12.0
pandas>=2.2.0
scipy>=1.12.0
matplotlib>=3.8.2
numpy>=1.26.0
librosa>=0.10.0